// src/enums/chat-message-external-statuses-enum.ts
export enum ChatMessageExternalStatusesEnum {
    QUEUED = 'queued',
    RINGING = 'ringing',
    IN_PROGRESS = 'in-progress',
    CANCELED = 'canceled',
    COMPLETED = 'completed',
    BUSY = 'busy',
    NO_ANSWER = 'no-answer',
    FAILED = 'failed',
    SENDING = 'sending',
    SENT = 'sent',
    DELIVERED = 'delivered',
    UNDELIVERED = 'undelivered',
    RECEIVING = 'receiving',
    RECEIVED = 'received',
    ACCEPTED = 'accepted',
    SCHEDULED = 'scheduled',
    READ = 'read',
    ERROR = 'error',
}